Add GZip and GZipDictionary compression models for benchmarking and comparison.  Also run valgrind against unit test and clean up various leaks and errant references.  Install valgrind and then run it this way:
valgrind --leak-check=full --show-reachable=yes --dsymutil=yes ./cpp-tests

As of this checkin the output indicates that the deflate library has
some "conditional based on uninitialized variable".  2 of them
Also no leaks now but we leave 4k reachable which seems to be the output
buffer from std::cout.  So the unit test is clean.